fix: move all interface to spec pkg
