+++ client.go
@@ -1,51 +1,22 @@
 package ibc
 
-// ClientType is a set of definitions of the data structures, initialisation logic,
-// validity predicate, and misbehaviour predicate required to operate a light client.
-type ClientType int
+import "github.com/DE-labtory/ibc/spec"
 
-// ClientState keeps arbitrary internal state to track verified roots and past misbehaviours.
-type ClientState struct {
-	frozen        bool
-	pastPublicKey map[string]bool
-	verifiedRoot  map[int]CommitmentRoot
+// Client is temporarily implementation of Client interface.
+type Client struct {
+	identifier     spec.Identifier
+	clientType     spec.ClientType
+	consensusState spec.ConsensusState
 }
 
-// Header provides information to update a ConsensusState.
-type Header struct {
-	sequence       int
-	commitmentRoot CommitmentRoot
-	signature      Signature
-	newPublicKey   PublicKey
-}
-
-// Evidence has candidate headers which may be misbehaviours.
-type Evidence struct {
-	h1 Header
-	h2 Header
-}
-
-type Client interface {
-	getIdentifier() Identifier
-	getClientType() ClientType
-	getConsensusState() ConsensusState
-}
-
-// ClientImpl is temporarily implementation of Client interface.
-type ClientImpl struct {
-	identifier     Identifier
-	clientType     ClientType
-	consensusState ConsensusState
-}
-
-func (c *ClientImpl) getIdentifier() Identifier {
+func (c *Client) getIdentifier() spec.Identifier {
 	return c.identifier
 }
 
-func (c *ClientImpl) getClientType() ClientType {
+func (c *Client) getClientType() spec.ClientType {
 	return c.clientType
 }
 
-func (c *ClientImpl) getConsensusState() ConsensusState {
+func (c *Client) getConsensusState() spec.ConsensusState {
 	return c.consensusState
 }
