+++ router.go
@@ -1,113 +1,69 @@
 package ibc
 
-// Router defines callback used by IBC handler.
-type Router interface {
-	onChanOpenInit(
-		order ChannelOrder,
-		connectionHops []Identifier,
-		portIdentifier Identifier,
-		channelIdentifier Identifier,
-		counterpartyPortIdentifier Identifier,
-		counterpartyChannelIdentifer Identifier,
-		version string)
+import "github.com/DE-labtory/ibc/spec"
 
-	onChanOpenTry(
-		order ChannelOrder,
-		connectionHops []Identifier,
-		portIdentifier Identifier,
-		channelIdentifier Identifier,
-		counterpartyPortIdentifier Identifier,
-		counterpartyChannelIdentifer Identifier,
-		version string)
-
-	onOpenAck(
-		portIdentifier Identifier,
-		channelIdentifier Identifier,
-		version string)
-
-	onChanOpenConfirm(
-		portIdentifier Identifier,
-		channelIdentifier Identifier)
-
-	onChanCloseInit(
-		portIdentifier Identifier,
-		channelIdentifier Identifier)
-
-	onChanCloseConfirm(
-		portIdentifier Identifier,
-		channelIdentifier Identifier)
-
-	onRecvPacket(packet interface{}) []byte
-
-	onTimeoutPacket(packet interface{})
-
-	onAcknowledgePacket(packet interface{})
-
-	onTimeoutPacketClose(packet interface{})
+type Router struct {
 }
 
-type RouterModule struct {
-}
-
-func (r *RouterModule) onChanOpenInit(
-	order ChannelOrder,
-	connectionHops []Identifier,
-	portIdentifier Identifier,
-	channelIdentifier Identifier,
-	counterpartyPortIdentifier Identifier,
-	counterpartyChannelIdentifer Identifier,
+func (r *Router) onChanOpenInit(
+	order spec.ChannelOrder,
+	connectionHops []spec.Identifier,
+	portIdentifier spec.Identifier,
+	channelIdentifier spec.Identifier,
+	counterpartyPortIdentifier spec.Identifier,
+	counterpartyChannelIdentifer spec.Identifier,
 	version string) {
 	panic("implement me")
 }
 
-func (r *RouterModule) onChanOpenTry(
-	order ChannelOrder,
-	connectionHops []Identifier,
-	portIdentifier Identifier,
-	channelIdentifier Identifier,
-	counterpartyPortIdentifier Identifier,
-	counterpartyChannelIdentifer Identifier,
+func (r *Router) onChanOpenTry(
+	order spec.ChannelOrder,
+	connectionHops []spec.Identifier,
+	portIdentifier spec.Identifier,
+	channelIdentifier spec.Identifier,
+	counterpartyPortIdentifier spec.Identifier,
+	counterpartyChannelIdentifer spec.Identifier,
 	version string) {
 	panic("implement me")
 }
 
-func (r *RouterModule) onOpenAck(
-	portIdentifier Identifier,
-	channelIdentifier Identifier,
+func (r *Router) onOpenAck(
+	portIdentifier spec.Identifier,
+	channelIdentifier spec.Identifier,
 	version string) {
 	panic("implement me")
 }
 
-func (r *RouterModule) onChanOpenConfirm(
-	portIdentifier Identifier,
-	channelIdentifier Identifier) {
+func (r *Router) onChanOpenConfirm(
+	portIdentifier spec.Identifier,
+	channelIdentifier spec.Identifier) {
 	panic("implement me")
 }
 
-func (r *RouterModule) onChanCloseInit(
-	portIdentifier Identifier,
-	channelIdentifier Identifier) {
+func (r *Router) onChanCloseInit(
+	portIdentifier spec.Identifier,
+	channelIdentifier spec.Identifier) {
 	panic("implement me")
 }
 
-func (r *RouterModule) onChanCloseConfirm(
-	portIdentifier Identifier,
-	channelIdentifier Identifier) {
+func (r *Router) onChanCloseConfirm(
+	portIdentifier spec.Identifier,
+	channelIdentifier spec.Identifier) {
 	panic("implement me")
 }
 
-func (r *RouterModule) onRecvPacket(packet interface{}) []byte {
+func (r *Router) onRecvPacket(packet interface{}) []byte {
 	panic("implement me")
 }
 
-func (r *RouterModule) onTimeoutPacket(packet interface{}) {
+func (r *Router) onTimeoutPacket(packet interface{}) {
 	panic("implement me")
 }
 
-func (r *RouterModule) onAcknowledgePacket(packet interface{}) {
+func (r *Router) onAcknowledgePacket(packet interface{}) {
 	panic("implement me")
 }
 
-func (r *RouterModule) onTimeoutPacketClose(packet interface{}) {
+func (r *Router) onTimeoutPacketClose(packet interface{}) {
 	panic("implement me")
 }
