+++ spec/blockchain.go
@@ -1,4 +1,4 @@
-package ibc
+package spec
 
 // Consensus is a Header generating function which takes the previous ConsensusState with the messages.
 type Consensus func(ConsensusState, []Message) Header
@@ -14,11 +14,6 @@ type Message interface{}
 // Blockchain is a consensus algorithm which generates valid headers.
 // It generates a unique list of headers starting from a genesis ConsensusState with arbitrary messages.
 type Blockchain interface {
-	getGenesis() Consensus
-	getConsensus() Consensus
-}
-
-type BlockchainImpl struct {
-	genesis   ConsensusState
-	consensus Consensus
+	GetGenesis() Consensus
+	GetConsensus() Consensus
 }
