+++ spec/machine.go
@@ -1,4 +1,4 @@
-package ibc
+package spec
 
 /*
 +----------------------------------------------------------------------------------+
