+++ spec/store.go
@@ -1,0 +1,7 @@
+package spec
+
+type KVStore interface {
+	get(path Path) Value
+	set(path Path, value Value)
+	delete(path Path)
+}
