+++ store.go
@@ -1,37 +1,33 @@
 package ibc
 
-type KVStore interface {
-	get(path Path) Value
-	set(path Path, value Value)
-	delete(path Path)
-}
+import "github.com/DE-labtory/ibc/spec"
 
 type ProvableStore struct {
 }
 
-func (*ProvableStore) get(path Path) Value {
+func (*ProvableStore) get(path spec.Path) spec.Value {
 	panic("implement me")
 }
 
-func (*ProvableStore) set(path Path, value Value) {
+func (*ProvableStore) set(path spec.Path, value spec.Value) {
 	panic("implement me")
 }
 
-func (*ProvableStore) delete(path Path) {
+func (*ProvableStore) delete(path spec.Path) {
 	panic("implement me")
 }
 
 type PrivateStore struct {
 }
 
-func (*PrivateStore) get(path Path) Value {
+func (*PrivateStore) get(path spec.Path) spec.Value {
 	panic("implement me")
 }
 
-func (*PrivateStore) set(path Path, value Value) {
+func (*PrivateStore) set(path spec.Path, value spec.Value) {
 	panic("implement me")
 }
 
-func (*PrivateStore) delete(path Path) {
+func (*PrivateStore) delete(path spec.Path) {
 	panic("implement me")
 }
